requirements.txt
openai
PyPDF2
pandas
pandas==2.2.3
peopledatalabs==4.2.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.38.0
loguru
regex
numpy
requests
pydantic
pdf2image
matplotlib
seaborn
fpdf
asts
kaleido
plotly